Split Repo Final Commit
Processing and C++ files separated into their own folders. A few file names were changed to better match the theoretical purpose of them. Readme should be the only incomplete file, however the code is self documenting and significant effort was put into its readability by anyone familiar or unfamiliar with C++.

The Processing files serve as a working visualization however they lack the readability of the C++ files. As a result I would recommend only using them to visualize the problem, as it was designed to do, and not to fix issues with the script.